(vim|bash)rc mods for NixOS (..)
Vim plugins are currently managed with vim-plug, but
will switch to nixpkgs packages, because for example
vim-fzf has a runtime dep on fzf source, and it is a
mess.

In order  for vim-fzf to  work, had to  download fzf
source (to dotfiles for  now), and hardcode the path
by adding it to rtp.